+++ Packages/src/Runtime/AnonymousAuthenticator.cs
@@ -1,0 +1,13 @@
+using System.Threading.Tasks;
+using Unity.Services.Authentication;
+
+namespace oojjrs.oauth
+{
+    public sealed class AnonymousAuthenticator : Authenticator
+    {
+        protected override async Task SignInAsync()
+        {
+            await AuthenticationService.Instance.SignInAnonymouslyAsync();
+        }
+    }
+}